% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approximate_correlation.R
\name{expected_target_product}
\alias{expected_target_product}
\title{Compute intermediate result for computation of autocorrelation between two
random variables in the target process.}
\usage{
expected_target_product(cdf, r, gamma)
}
\arguments{
\item{cdf}{Target distribution's Cumulative Distribution Function. All of its
parameters need to be set, so that an element from the value range is the
only argument to provided.}

\item{r}{Autocorrelation between two variables of the base process. Should be
selected such that it induces the targeted autocorrelation in the target
process.}

\item{gamma}{Controls the quality of the integral approximation.
Computation stops once the next subsum of integral plateaus being computed
has a relative size compared to the current total sum smaller then the
gamma.}
}
\value{
Expected value of product of two random variables in the target
process, which are the product of the inverse transform of two random
variables in the base process with autocorrelation r
}
\description{
Compute intermediate result for computation of autocorrelation between two
random variables in the target process.
}
\details{
This function approximates the expected value of the product of two random
variables in the target process. By making use of the fact that the function
being approximated is piecewise steady, it sums up  the exact values of
integrals over plateaus, where a solution can be found analytically, until a
threshold is reached.
}
