% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DARTA.R
\name{find_r_interpol}
\alias{find_r_interpol}
\title{Compute a fitting autocorrelation structure for the base process such that
it approximates the provided target process, by use of a interpolation method.}
\usage{
find_r_interpol(
  cdf,
  mean,
  var,
  rho,
  gamma,
  n_interpol,
  poly_deg = 9,
  use_caching = T,
  cdf_name_parameterized
)
}
\arguments{
\item{cdf}{Marginal CDF of target process.}

\item{mean}{Mean of marginal CDF of target process.}

\item{var}{Variance of marginal CDF of target process.}

\item{rho}{Target autocorrelation Structure of target process, provided as a
vector of numerics. Position on the vector describes the lag for which the
numeric autocorrelation value is supposed to occur.}

\item{gamma}{Threshold for computation of the expected value of the
product of two random variables of the target process.}

\item{n_interpol}{Controls the number of equidistant points in the base autocorrelation space for which resulting autocorrelations in the target autocorrelation space are computed, in order to fit a polynomial through them.}

\item{poly_deg}{Sets the degree of the polynomial fit through 'n_interpol' many points.}

\item{use_caching}{Should cached values be used, and saved after computation, or not.}

\item{cdf_name_parameterized}{Name of the target Distribution, together with
its parameters, separated by dashes. Used for naming the cache-file, which
stores a map of values from the base process autocorrelation space to the
target process autocorrelation space.}
}
\value{
A vector containing a fitting autocorrelation structure for the base
process.
}
\description{
Compute a fitting autocorrelation structure for the base process such that
it approximates the provided target process, by use of a interpolation method.
}
\details{
A map of base process autocorrelation values to target process
autocorrelation values is stored in the \code{.interpol_caches} file in the
current working directory, if the 'use_caching' parameter is set to TRUE.
Marginal distributions that assign significant
probability mass to a large amount of their value range tend to require
longer computations in order to find a fitting base autocorrelation
structure, and caching removes the need for this computation on repeated
generation attempts.
}
