% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_distributions.R
\name{generate_binomial}
\alias{generate_binomial}
\title{Generate autocorrelated time-series with marginal binomial distribution.}
\usage{
generate_binomial(
  n,
  size,
  prob,
  rho,
  gamma = 1e-05,
  method = "interpol",
  n_interpol = 30,
  epsilon = 0.001,
  use_caching = T
)
}
\arguments{
\item{n}{length of time-series to be generated, i.e., total size of the
vector of random numbers that is returned.}

\item{size}{Size parameter of the binomial distribution.}

\item{prob}{Probability parameter of the binomial distribution.}

\item{rho}{Autocorrelation structure to be approximated, as a vector of
\code{numerics}.}

\item{gamma}{Numeric, controls accuracy of DARTA's autocorrelation
approximation mechanism, lower being more accurate. Defaults to 10^-5.}

\item{method}{Defines method for autocorrelation structure approximation.
'binary' employs a binary search, ending when autocorrelation in target
process is within error margin 'epsilon' of desired autocorrelation.
'interpol' fits a function through 'n_interpol' autocorrelation values
calculated from equidistant points in the base process autocorrelation
space, then interpolates a fitting base process autocorrelation from these
values.}

\item{n_interpol}{Natural number, decides how many equidistant points in the
base process autocorrelation space should be sampled in order to fit a
polynomial function. Defaults to 30.}

\item{epsilon}{Numeric between 0 and 1, controls acceptable error within
which the target autocorrelation is to be approximated by the base process
when the 'binary' method is used. Defaults to 0.001.}

\item{use_caching}{Logical, deciding whether results should be cached, and
cached results used in the computation. Defaults to TRUE.}
}
\value{
Vector containing a time-series of length n with autocorrelation rho
and binomial marginal distribution, or NULL if base process is not
stationary.
}
\description{
Generate autocorrelated time-series with marginal binomial distribution.
}
\examples{
library(DARTA)
# define distribution parameters
size <- 30
prob <- 0.7

# define autocorrelation structure as a vector, starting with lag 1:
rho <- c(0.7,0.4,0.1)

# generate sample
vec <- generate_binomial(n = 1e6, size =size, prob=prob, rho=rho, method = "interpol", use_caching = F)

# plot autocorrelation
acf(vec)

# plot empirical CDF
plot(ecdf(vec))

# compare to binomial distribution
plot(ecdf(rbinom(1e5,size,prob)))

}
\seealso{
\code{\link{generate_distribution}}
}
