% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_distributions.R
\name{generate_custom_distribution}
\alias{generate_custom_distribution}
\title{Generate time-series of a custom distribution}
\usage{
generate_custom_distribution(
  n,
  cdf,
  inv,
  mean,
  name,
  var,
  rho,
  gamma = 1e-05,
  method = "interpol",
  n_interpol = 30,
  epsilon = 0.001,
  use_caching = T,
  ...
)
}
\arguments{
\item{n}{length of time-series to be generated, i.e., total size of the
vector of random numbers that is returned.}

\item{cdf}{Cumulative distribution function of the custom distribion}

\item{inv}{Inverse cumulative distribution function, i.e. qunatile function, of cumstom distribution}

\item{mean}{Mean of Custom distribution}

\item{name}{Name of the custom distribution, used for caching. Should be chosen uniquely for each cdf, but different parameterizations can be cached for the same name.}

\item{var}{Variance of custom distribution}

\item{rho}{Autocorrelation structure to be approximated, as a vector of
\code{numerics}.}

\item{gamma}{Numeric, controls accuracy of DARTA's autocorrelation
approximation mechanism, lower being more accurate. Defaults to 10^-5.
product of two random variables of the target process. defaults to 10^-5.}

\item{method}{Defines method for autocorrelation structure approximation.
'binary' employs a binary search, ending when autocorrelation in target
process is within error margin 'epsilon' of desired autocorrelation.
'interpol' fits a function through 'n_interpol' autocorrelation values
calculated fdevrom equidistant points in the base process autocorrelation
space, then interpolates a fitting base process autocorrelation from these
values.}

\item{n_interpol}{Natural number, decides how many equidistant points in the
base process autocorrelation space should be sampled in order to fit a
polynomial function. Defaults to 30.}

\item{epsilon}{Numeric between 0 and 1, controls acceptable error within
which the target autocorrelation is to be approximated by the base process
when the 'binary' method is used. Defaults to 0.001.}

\item{use_caching}{Logical, deciding whether results should be cached, and
cached results used in the computation. Defaults to TRUE.}
}
\value{
Vector containing a time-series of length n with autocorrelation rho
and cumstom marginal distribution.
}
\description{
Generate time-series of a custom distribution
}
\examples{
library(DARTA)
# define distribution parameters
min <- 50
max <- 150

# define autocorrelation structure as a vector, starting with lag 1:
rho <- c(0.3,0.2,0.3,0.2,0.3,0.2)

# generate sample
vec <- generate_uniform(n = 1e5, min = min, max = max, rho=rho, method = "interpol", use_caching = F)

# plot autocorrelation
acf(vec)

# plot empirical CDF
plot(ecdf(vec))

# compare to uniform distribution
plot(ecdf(runif(1e5,min, max)))

}
\seealso{
\code{\link{generate_distribution}}
}
