% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_distributions.R
\name{generate_distribution}
\alias{generate_distribution}
\title{Generate time-series of the provided marginal distribution and
autocorrelation structure.}
\usage{
generate_distribution(
  n,
  rho,
  distribution_name = c("nbinomial", "binomial", "poisson", "uniform", "custom"),
  gamma = 1e-05,
  method = c("interpol", "binary"),
  epsilon = 0.001,
  n_interpol = 30,
  use_caching = T,
  ...
)
}
\arguments{
\item{n}{length of time-series to be generated, i.e., total size of the
vector of random numbers that is returned.}

\item{rho}{Autocorrelation structure to be approximated, as a vector of
numerics.}

\item{distribution_name}{Name of the Distribution to be generated. Currently
supported are "nbinomial", "binomial", "poisson", "uniform", and "custom".}

\item{gamma}{Numeric, controls accuracy of DARTA's autocorrelation
approximation mechanism, lower being more accurate. Defaults to 10^-5.
product of two random variables of the target process.}

\item{method}{Defines method for autocorrelation structure approximation.
'binary' employs a binary search, ending when autocorrelation in target
process is within error margin 'epsilon' of desired autocorrelation.
'interpol' fits a function through 'n_interpol' autocorrelation values
calculated from equidistant points in the base process autocorrelation
space, then interpolates a fitting base process autocorrelation from these
values.}

\item{epsilon}{Numeric between 0 and 1, controls acceptable error within
which the target autocorrelation is to be approximated by the base process
when the 'binary' method is used. Defaults to 0.001.}

\item{n_interpol}{Natural number, decides how many equidistant points in the
base process autocorrelation space should be sampled in order to fit a
polynomial function. Defaults to 30.}

\item{use_caching}{Logical, deciding whether results should be cached, and
cached results used in the computation.}
}
\value{
Vector containing a time-series of length n with autocorrelation rho
and marginal distribution defined by distribution_name, or NULL if base
process is not stationary.
}
\description{
Generate time-series of the provided marginal distribution and
autocorrelation structure.
}
\examples{
generate_DARTA(n = 10000, rho = c(0.5,0.4,0.3), distribution_name = "nbinomial", size = 10, prob = 0.6, method = "interpol")
}
\seealso{
\code{\link{generate_DARTA}}
}
