% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_distributions.R
\name{generate_nbinomial}
\alias{generate_nbinomial}
\title{Generate autocorrelated time-series with marginal negative binomial distribution.}
\usage{
generate_nbinomial(
  n,
  size,
  prob,
  rho,
  gamma = 1e-05,
  method = "interpol",
  n_interpol = 20,
  epsilon = 0.001,
  use_caching = T
)
}
\arguments{
\item{n}{length of time-series to be generated, i.e., total size of the
vector of random numbers that is returned.}

\item{size}{Size parameter of the Negative Binomial distribution.}

\item{prob}{Probability parameter of the Negative Binomial distribution.}

\item{rho}{Autocorrelation structure to be approximated, as a vector of
\code{numerics}.}

\item{gamma}{Controls accuracy of DARTA's autocorrelation approximation mechanism, lower being more accurate. Defaults to 10^-5.}

\item{method}{Defines method for autocorrelation structure approximation.
'binary' employs a binary search, ending when autocorrelation in target
process is within error margin 'epsilon' of desired autocorrelation.
'interpol' fits a function through 'n_interpol' autocorrelation values
calculated from equidistant points in the base process autocorrelation
space, then interpolates a fitting base process autocorrelation from these
values.}

\item{epsilon}{Controls acceptable error within which the target
autocorrelation is to be approximated by the base process when the 'binary' method is used. Defaults to
0.001.}
}
\value{
Vector containing a time-series of length n with autocorrelation rho
and Negative Binomial marginal distribution, or NULL if base
process is not stationary.
}
\description{
Generate autocorrelated time-series with marginal negative binomial distribution.
}
\examples{
library(DARTA)
# define distribution parameters
size <- 10
prob <- 0.9

# define autocorrelation structure as a vector, starting with lag 1:
rho <- c(0.5,0.4,0.3)

# generate sample
vec <- generate_nbinomial(n = 1e5, size =size, prob=prob, rho=rho, method = "interpol", use_caching = F)

# plot autocorrelation
acf(vec)

# plot empirical CDF
plot(ecdf(vec))

# compare to negative bin distribution
plot(ecdf(rnbinom(1e5,size,prob)))



}
\seealso{
\code{\link{generate_distribution}}
}
