% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_distributions.R
\name{generate_poisson}
\alias{generate_poisson}
\title{Generate autocorrelated time-series with marginal poisson distribution.}
\usage{
generate_poisson(
  n,
  lambda,
  rho,
  gamma = 1e-05,
  method = "interpol",
  n_interpol = 20,
  epsilon = 0.001,
  use_caching = T
)
}
\arguments{
\item{n}{length of time-series to be generated, i.e., total size of the
vector of random numbers that is returned.}

\item{lambda}{Lambda parameter of the Poisson distribution.}

\item{rho}{Autocorrelation structure to be approximated, as a vector of
\code{numerics}.}

\item{gamma}{Controls accuracy of DARTA's autocorrelation approximation mechanism, lower being more accurate. Defaults to 10^-5.}

\item{method}{Defines method for autocorrelation structure approximation.
'binary' employs a binary search, ending when autocorrelation in target
process is within error margin 'epsilon' of desired autocorrelation.
'interpol' fits a function through 'n_interpol' autocorrelation values
calculated from equidistant points in the base process autocorrelation
space, then interpolates a fitting base process autocorrelation from these
values.}

\item{epsilon}{Controls acceptable error within which the target
autocorrelation is to be approximated by the base process when the 'binary' method is used. Defaults to
0.001.}
}
\value{
Vector containing a time-series of length n with autocorrelation rho
and marginal Poisson-distribution, or NULL if base
process is not stationary.
}
\description{
Generate autocorrelated time-series with marginal poisson distribution.
}
\seealso{
\code{\link{generate_distribution}}
}
