% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DARTA.R
\name{get_correlation_bound}
\alias{get_correlation_bound}
\title{Compute the minimum and maximum possible values for autocorrelation between
variables of the target process. If any target autocorrelation present in the
target autocorrelation structures lies outside of the computed bounds, the
process cannot produce valid results.}
\usage{
get_correlation_bound(cdf, mean, var, gamma)
}
\arguments{
\item{cdf}{Target distribution's Cumulative Distribution Function. All of its
parameters need to be set, so that an element from the value range is the
only argument to provided.}

\item{mean}{Mean of the target distribution.}

\item{var}{Variance of the target distribution.}

\item{gamma}{Defines a threshhold for the relative size of the subsummand in
the computation of function \code{\link{expected_target_product}}}
}
\description{
Compute the minimum and maximum possible values for autocorrelation between
variables of the target process. If any target autocorrelation present in the
target autocorrelation structures lies outside of the computed bounds, the
process cannot produce valid results.
}
