% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_r_interpol.R
\name{get_poly_target}
\alias{get_poly_target}
\title{Compute values to which to fit the polynomial, without caching}
\usage{
get_poly_target(fitting_base, cdf, gamma, mean, var)
}
\description{
Compute values to which to fit the polynomial, without caching
}
