% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DARTA.R
\name{get_poly_target_cached}
\alias{get_poly_target_cached}
\title{Compute values to which to fit the polynomial, with caching enabled}
\usage{
get_poly_target_cached(cache_file, fitting_base, cdf, gamma, mean, var)
}
\description{
Compute values to which to fit the polynomial, with caching enabled
}
