% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DARTA.R
\name{get_target_val_cached}
\alias{get_target_val_cached}
\title{Compute single autocorrelation value in target autocorrelation space from value in base autocorrelation space}
\usage{
get_target_val_cached(base_val, cache, cdf, gamma, mean, var)
}
\description{
Compute single autocorrelation value in target autocorrelation space from value in base autocorrelation space
}
