% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approximate_correlation.R
\name{integrand_plateau}
\alias{integrand_plateau}
\title{Compute the value of the integral for a plateau on the integrand function
surface.}
\usage{
integrand_plateau(j, k, cdf, r)
}
\arguments{
\item{j}{Element of target distribution's value range.}

\item{k}{Element of target distribution's value range.}

\item{cdf}{Target distribution's Cumulative Distribution Function.
Needs to be fully parameterized.}

\item{r}{Autocorrelation between two variables of the base process.}
}
\description{
Compute the value of the integral for a plateau on the integrand function
surface.
}
\seealso{
\code{\link{VGAM::pbinorm}}
}
