% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approximate_correlation.R
\name{integrand_plateau_symmetric}
\alias{integrand_plateau_symmetric}
\title{Compute the value of the integral for a plateau on the integrand function
surface. Plateaus edges have to be defined by the same value from the target
distribution's value range.}
\usage{
integrand_plateau_symmetric(k, cdf, r)
}
\arguments{
\item{k}{Element of target distribution's value range.}

\item{cdf}{Target distribution's Cumulative Distribution Function.
Needs to be fully parameterized.}

\item{r}{Autocorrelation between two variables of the base process.}
}
\description{
Compute the value of the integral for a plateau on the integrand function
surface. Plateaus edges have to be defined by the same value from the target
distribution's value range.
}
\seealso{
\code{\link{VGAM::pbinorm}}
}
