% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DARTA.R
\name{is_stationary}
\alias{is_stationary}
\title{Check if distribution resulting from alpha is stationary.}
\usage{
is_stationary(alpha)
}
\arguments{
\item{alpha}{Vector of \code{numerics} defining the coefficients in a
recursively defined stochastic process.}
}
\value{
Logical indicating if the resulting distribution is stationary or not.
}
\description{
Check if distribution resulting from alpha is stationary.
}
