% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DARTA.R
\name{expected_target_product}
\alias{expected_target_product}
\title{Computes an approximation of the integral over the integral function. This is
done by summing up the exact values of integrals over plateaus, until a
threshold is reached.}
\usage{
expected_target_product(cdf, r, gamma)
}
\arguments{
\item{cdf}{Target distribution's Cumulative Distribution Function. All of its
parameters need to be set, so that an element from the value range is the
only argument to provided.}

\item{r}{Autocorrelation between two variables of the base process. Should be
selected such that it induces the targeted autocorrelation in the target
process.}

\item{gamma}{Controls the quality of the integral approximation.
Computation stops once the next subsum of integral plateaus being computed
has a relative size compared to the current total sum smaller then the
gamma.}
}
\description{
Computes an approximation of the integral over the integral function. This is
done by summing up the exact values of integrals over plateaus, until a
threshold is reached.
}
