% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_distributions.R
\name{generate_distribution}
\alias{generate_distribution}
\title{Generate time-series of the provided marginal distribution and
autocorrelation structure.}
\usage{
generate_distribution(
  n,
  rho,
  distribution_name = c("nbinomial", "binomial", "poisson", "uniform"),
  gamma = 1e-05,
  method = c("interpol", "binary"),
  epsilon = 0.001,
  n_interpol = 20,
  use_caching = T,
  ...
)
}
\arguments{
\item{n}{length of time-series to be generated, i.e., total size of the
vector of random numbers that is returned.}

\item{rho}{Autocorrelation structure to be approximated, as a vector of
numerics.}

\item{distribution_name}{Name of the Distribution to be generated. Currently
supported are "nbinomial", "binomial", "poisson", and "uniform".}

\item{gamma}{Threshold for computation of the expected value of the
product of two random variables of the target process. defaults to 10^-5.}

\item{epsilon}{Controls acceptable error within which the target
autocorrelation is to be approximated by the base process. Defaults to
0.001.}
}
\value{
Vector containing a time-series of length n with autocorrelation rho
and marginal distribution defined by distribution_name, or NULL if base
process is not stationary.
}
\description{
Generate time-series of the provided marginal distribution and
autocorrelation structure.
}
\examples{
generate_DARTA(n = 10000, rho = c(0.7,0.5,0.3), distribution_name = "nbinomial", param1 = 10, param2 = 0.6)
}
\seealso{
\code{\link{generate_DARTA}}
}
