% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_distributions.R
\name{generate_poisson}
\alias{generate_poisson}
\title{Generate autocorrelated time-series with marginal poisson distribution.}
\usage{
generate_poisson(
  n,
  lambda,
  rho,
  gamma = 1e-05,
  method = "interpol",
  n_interpol = 30,
  epsilon = 0.001,
  use_caching = T
)
}
\arguments{
\item{n}{length of time-series to be generated, i.e., total size of the
vector of random numbers that is returned.}

\item{lambda}{Lambda parameter of the Poisson distribution.}

\item{rho}{Autocorrelation structure to be approximated, as a vector of
\code{numerics}.}

\item{gamma}{Numeric, controls accuracy of DARTA's autocorrelation
approximation mechanism, lower being more accurate. Defaults to 10^-5.
product of two random variables of the target process. defaults to 10^-5.}

\item{method}{Defines method for autocorrelation structure approximation.
'binary' employs a binary search, ending when autocorrelation in target
process is within error margin 'epsilon' of desired autocorrelation.
'interpol' fits a function through 'n_interpol' autocorrelation values
calculated from equidistant points in the base process autocorrelation
space, then interpolates a fitting base process autocorrelation from these
values.}

\item{n_interpol}{Natural number, decides how many equidistant points in the
base process autocorrelation space should be sampled in order to fit a
polynomial function. Defaults to 30.}

\item{epsilon}{Numeric between 0 and 1, controls acceptable error within
which the target autocorrelation is to be approximated by the base process
when the 'binary' method is used. Defaults to 0.001.}

\item{use_caching}{Logical, deciding wether results should be cached, and
cached results used in the computation. Defaults to TRUE.}
}
\value{
Vector containing a time-series of length n with autocorrelation rho
and marginal poisson distribution, or NULL if base process is not
stationary.
}
\description{
Generate autocorrelated time-series with marginal poisson distribution.
}
\examples{
library(DARTA)
# define distribution parameter
lambda <- 1.5

# define autocorrelation structure as a vector, starting with lag 1:
rho <- c(0.9)

# generate sample
vec <- generate_poisson(n = 1e5, lambda = lambda, rho=rho, method = "interpol", use_caching = F)

# plot autocorrelation
acf(vec)

# plot empirical CDF
plot(ecdf(vec))

# compare to poisson distribution
plot(ecdf(rpois(1e5,lambda)))
}
\seealso{
\code{\link{generate_distribution}}
}
