% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DARTA.R
\name{get_gamma}
\alias{get_gamma}
\title{Build Gamma matrix for solving Yule-Walker-Equations.}
\usage{
get_gamma(r)
}
\arguments{
\item{r}{Vector of \code{numerics} describing the autocorrelation structure
of the base process.}
}
\description{
Build Gamma matrix for solving Yule-Walker-Equations.
}
