% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DARTA.R
\name{get_target_correlation}
\alias{get_target_correlation}
\title{Helper function for computing the autocorrelation between to random variables
of the target process based on the properties of the corresponding random
variables of the currently evaluated base process, i.e., the selected base
process autocorrelation and target process CDF.}
\usage{
get_target_correlation(expected_target_product, mean, var)
}
\arguments{
\item{expected_target_product}{Approximation (or exact value, if possible) of
the integral function defining the expected value of two random variables of
the target process. Usually expects the return value of
\link{expected_target_product}.}

\item{mean}{Mean of the target distribution.}

\item{var}{Variance of the target distribution.}
}
\description{
Helper function for computing the autocorrelation between to random variables
of the target process based on the properties of the corresponding random
variables of the currently evaluated base process, i.e., the selected base
process autocorrelation and target process CDF.
}
