% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DARTA.R
\name{integrand_plateau}
\alias{integrand_plateau}
\title{Computes the value of the integral for a plateau on the integrand function
surface. Plateau is specified by the values of the target distribution value
range that define its edges.}
\usage{
integrand_plateau(j, k, cdf, r)
}
\arguments{
\item{j}{Element of target distribution's value range.}

\item{k}{Element of target distribution's value range.}

\item{cdf}{Target distribution's Cumulative Distribution Function. Needs to be fully parameterized.}

\item{r}{Autocorrelation between two variables of the base process. Should be
selected such that it induces the targeted autocorrelation in the target
process.}
}
\description{
Computes the value of the integral for a plateau on the integrand function
surface. Plateau is specified by the values of the target distribution value
range that define its edges.
}
\seealso{
\code{\link{VGAM::pbinorm}}
}
