% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DARTA.R
\name{integrand_plateau_symmetric}
\alias{integrand_plateau_symmetric}
\title{Computes the value of the integral for a plateau on the integrand function
surface. Plateaus edges have to be defined by the same value from the target
distribution's value range.}
\usage{
integrand_plateau_symmetric(k, cdf, r)
}
\arguments{
\item{k}{Element of target distribution's value range.}

\item{cdf}{Target distribution's Cumulative Distribution Function. All of its
parameters need to be set, so that an element from the value range is the
only argument to provided.}

\item{r}{Autocorrelation between two variables of the base process. Should be
selected such that it induces the targeted autocorrelation in the target
process.}
}
\description{
Computes the value of the integral for a plateau on the integrand function
surface. Plateaus edges have to be defined by the same value from the target
distribution's value range.
}
